feat: implement encrypted form handling and refactor backup restore logic
